+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 FEATURES:
 - Added `cidrcontains` function to check if a CIDR contains an IP address.
 - Added `urldecode` function to decode a URL-encoded string.
+- Added `base64gunzip` function to decode a base64-encoded string and then decompress it using gzip.